15 Add support for AutoFlush events
